Move NonFungibleData derive to radix-engine-derive
