+++ radix-engine-derive/src/lib.rs
@@ -3,6 +3,7 @@ mod decode;
 mod describe;
 mod encode;
 mod legacy_describe;
+mod non_fungible_data;
 mod utils;
 
 use proc_macro::TokenStream;
@@ -98,6 +99,27 @@ pub fn categorize(input: TokenStream) -> TokenStream {
 #[proc_macro_derive(ScryptoDescribe, attributes(sbor))]
 pub fn describe(input: TokenStream) -> TokenStream {
     describe::handle_describe(proc_macro2::TokenStream::from(input))
+        .unwrap_or_else(|err| err.to_compile_error())
+        .into()
+}
+
+/// Derive code that describe a non-fungible data structure.
+///
+/// # Example
+///
+/// ```ignore
+/// use scrypto::prelude::*;
+///
+/// #[derive(NonFungibleData)]
+/// pub struct MyStruct {
+///     pub field_1: u32,
+///     #[scrypto(mutable)]
+///     pub field_2: String,
+/// }
+/// ```
+#[proc_macro_derive(NonFungibleData, attributes(scrypto))]
+pub fn non_fungible_data(input: TokenStream) -> TokenStream {
+    non_fungible_data::handle_non_fungible_data(proc_macro2::TokenStream::from(input))
         .unwrap_or_else(|err| err.to_compile_error())
         .into()
 }
