+++ radix-engine-derive/src/utils.rs
@@ -78,3 +78,8 @@ pub fn is_describing_skipped(f: &Field) -> bool {
     let parsed = extract_attributes(&f.attrs);
     parsed.contains_key("skip") || parsed.contains_key("skip_describing")
 }
+
+pub fn is_mutable(f: &Field) -> bool {
+    let parsed = extract_attributes(&f.attrs);
+    parsed.contains_key("mutable")
+}
