+++ scrypto-derive/src/utils.rs
@@ -1,15 +1,6 @@
-use std::collections::HashMap;
 use std::io::Write;
 use std::process::Command;
 use std::process::Stdio;
-
-use syn::punctuated::Punctuated;
-use syn::token::Comma;
-use syn::Attribute;
-use syn::Expr;
-use syn::ExprLit;
-use syn::Field;
-use syn::Lit;
 
 #[allow(dead_code)]
 pub fn print_generated_code<S: ToString>(kind: &str, code: S) {
@@ -37,44 +28,4 @@ pub fn print_generated_code<S: ToString>(kind: &str, code: S) {
             }
         }
     }
-}
-
-pub fn extract_attributes(attrs: &[Attribute]) -> HashMap<String, Option<String>> {
-    let mut configs = HashMap::new();
-
-    for attr in attrs {
-        if !attr.path.is_ident("scrypto") {
-            continue;
-        }
-
-        if let Ok(parsed) = attr.parse_args_with(Punctuated::<Expr, Comma>::parse_terminated) {
-            parsed.into_iter().for_each(|s| match s {
-                Expr::Assign(assign) => {
-                    if let Expr::Path(path_expr) = assign.left.as_ref() {
-                        if let Some(ident) = path_expr.path.get_ident() {
-                            if let Expr::Lit(ExprLit {
-                                lit: Lit::Str(s), ..
-                            }) = assign.right.as_ref()
-                            {
-                                configs.insert(ident.to_string(), Some(s.value()));
-                            }
-                        }
-                    }
-                }
-                Expr::Path(path_expr) => {
-                    if let Some(ident) = path_expr.path.get_ident() {
-                        configs.insert(ident.to_string(), None);
-                    }
-                }
-                _ => {}
-            })
-        }
-    }
-
-    configs
-}
-
-pub fn is_mutable(f: &Field) -> bool {
-    let parsed = extract_attributes(&f.attrs);
-    parsed.contains_key("mutable")
 }
