+++ scrypto/src/lib.rs
@@ -56,11 +56,13 @@ pub use macros::*;
 
 // Re-export radix engine derives
 pub extern crate radix_engine_derive;
-pub use radix_engine_derive::{LegacyDescribe, ScryptoCategorize, ScryptoDecode, ScryptoEncode};
+pub use radix_engine_derive::{
+    LegacyDescribe, NonFungibleData, ScryptoCategorize, ScryptoDecode, ScryptoEncode,
+};
 
 // Re-export Scrypto derive.
 extern crate scrypto_derive;
-pub use scrypto_derive::{blueprint, import, NonFungibleData};
+pub use scrypto_derive::{blueprint, import};
 
 pub extern crate radix_engine_interface;
 pub extern crate scrypto_abi;
